+++ tests/testthat/test-durations.R
@@ -91,6 +91,14 @@ test_that("parsing with 0 units works as expected", {
   expect_equal(period("y 0m 2d 0H 0M 1s"), years(1) + days(2) + seconds(1))
 })
 
+test_that("as.integer works with durrations", {
+  d1 <- duration("2d 0H 2M 1s")
+  expect_equal(as.integer(d1, unit = "minute"), as.integer(as.numeric(d1, unit = "minute")))
+  expect_equal(as.integer(d1, unit = "hour"), as.integer(as.numeric(d1, unit = "hour")))
+  expect_equal(as.integer(d1), as.numeric(d1))
+  expect_type(as.integer(d1), "integer")
+})
+
 test_that("make_difftime handles vectors", {
   x <- as.POSIXct(c("2008-08-03 13:01:59", "2008-08-03 13:01:59"), tz = "UTC")
   y <- difftime(x + c(