+++ tests/testthat/test-periods.R
@@ -559,6 +559,14 @@ test_that("c.Period doesn't fail with empty elements", {
   )
 })
 
+test_that("as.integer works with periods", {
+  p <- period("2d 0H 2M 1s")
+  expect_equal(as.integer(p, unit = "minute"), as.integer(as.numeric(p, unit = "minute")))
+  expect_equal(as.integer(p, unit = "hour"), as.integer(as.numeric(p, unit = "hour")))
+  expect_equal(as.integer(p), as.numeric(p))
+  expect_type(as.integer(p), "integer")
+})
+
 test_that("c.Period works with named elements", {
   nmd <- c(a = hours(1), b = minutes(30))
   umd <- c(hours(1), minutes(30))